Fix completion of (transparent) blockwise transfers.
When a payload is too large to be transferred in a single message, the
BlockwiseLayer transparently replaces the request/response with a
sequence of exchanges for transferring the chunks of payload data. At
the end the data is re-assembled into a single request/response and
handed over to the application layer.

In order to correctly clean up the state kept for the sequence of
message exchanges the original request needs to get assigned the token
used for the sequence of exchanges so that it can be completed in the
matcher correctly.